refactor: reorganize imports and improve module structure for clarity; resolve circular import issues
