+++ src/handlers/employees.py
@@ -8,12 +8,9 @@ from telebot import types
 from src.config import bot, process_in_progress, edit_employee_data, authorized_ids, add_employee_data, add_keyword_data
 from src.database import DatabaseConnection, update_authorized_users
 from src.handlers import authorized_only
-from src.handlers.authorization import authorized_only
-from src.integrations import proceed_find_user_id, update_employee_in_crm, delete_employee_from_crm, \
-    remove_user_from_chat
-from src.integrations.crm_api_functions import add_employee_to_crm
-from src.integrations.telethon_functions import proceed_find_user_id
-from src.utils import button_names
+from src.integrations.crm_api_functions import add_employee_to_crm, delete_employee_from_crm, update_employee_in_crm
+from src.integrations.telethon_functions import proceed_find_user_id, remove_user_from_chat
+from src.utils.main_menu_buttons import button_names
 
 
 @bot.callback_query_handler(func=lambda call: call.data.startswith('add_employee_'))
